apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-platform-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "sqlalchemy==2.0.35",
  "alembic==1.13.2",
  "psycopg[binary]==3.2.1",
  "pydantic-settings==2.4.0",
  "python-jose[cryptography]==3.3.0",
  "passlib[argon2]==1.7.4",
  "python-multipart==0.0.9",
  "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "httpx==0.27.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
